+++ inst/models/shiny/shinydashboard_ga4.R
@@ -5,10 +5,10 @@ library(googleAnalyticsR)
 
 gar_set_client(web_json = "{{ web_json }}",
                scopes = "{{ scopes }}")
+options(googleAuthR.redirect = "{{ deployed_url }}")
 
 # loads a pre-existing model
 model <- ga_model_load("{{ ga_model }}")
-
 modelUi <- model$shiny_module$ui
 modelServer <- model$shiny_module$server