+++ inst/models/shiny/template_ga4.R
@@ -4,6 +4,7 @@ library(googleAnalyticsR)
 
 gar_set_client(web_json = "{{ web_json }}",
                scopes = "{{ scopes }}")
+options(googleAuthR.redirect = "{{ deployed_url }}")
 
 # loads a pre-existing model
 model <- ga_model_load("{{ ga_model }}")